+++ lib/ayadn/api.rb
@@ -93,6 +93,10 @@ module Ayadn
       get_parsed_response(Endpoints.new.search_users(words, options))
     end
 
+    def search_annotations words, options
+      get_parsed_response(Endpoints.new.search_annotations(words, options))
+    end
+
     def get_followings(username)
       build_list(username, :followings)
     end