+++ lib/ayadn/endpoints.rb
@@ -181,6 +181,11 @@ module Ayadn
       "#{@posts_url}search?annotation_types=#{anno}&access_token=#{Settings.user_token}#{@options_list}"
     end
 
+    def search_channels words, options
+        @options_list = API.build_query({count: Settings.options[:counts][:search]})
+      "#{@channels_url}search?q=#{words}&order=popularity&access_token=#{Settings.user_token}#{@options_list}"
+    end
+
     def user(username)
       "#{@users_url}#{username}?access_token=#{Settings.user_token}&include_user_annotations=1"
     end